Safari-611.1.10/LayoutTests/fast/dom/Element/scrolling-funtions-on-body-quirks-expected.txt

Tests scrollTo/By(x,y) and scrollTo/By(scrollOptions) on the body in quirks mode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS scroller.scrollLeft is 20
PASS scroller.scrollTop is 500
PASS window.scrollX is 20
PASS window.scrollY is 500
PASS scroller.scrollLeft is 70
PASS scroller.scrollTop is 560
PASS window.scrollX is 70
PASS window.scrollY is 560
PASS scroller.scrollLeft is 100
PASS scroller.scrollTop is 400
PASS window.scrollX is 100
PASS window.scrollY is 400
PASS successfullyParsed is true

TEST COMPLETE
